% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{address_to_bytes}
\alias{address_to_bytes}
\title{Convert an address to 2 bytes (lo_byte, hi_byte) as is the 6502 way}
\usage{
address_to_bytes(address)
}
\arguments{
\item{address}{address from $0000 to $ffff}
}
\value{
integer vector of length 2 with (lo_byte, hi_byte)
}
\description{
Convert an address to 2 bytes (lo_byte, hi_byte) as is the 6502 way
}
