% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile.R
\name{compile}
\alias{compile}
\title{Compile ASM to prg_df}
\usage{
compile(asm)
}
\arguments{
\item{asm}{single character string containing all ASM (including newlines)}
}
\value{
prg_df main assembler data.frame
}
\description{
All the intermediate assembler steps are still works-in-progress and can
be quite fragile.
}
\details{
This function is only recommneded if you're sure that the code compiles OK,
otherwise tracking down errors is going to be pretty tough
}
