% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{create_prg_df_row}
\alias{create_prg_df_row}
\title{Determine the opcode, address mode etc for a sequence of tokens representing a single instruction}
\usage{
create_prg_df_row(tokens)
}
\arguments{
\item{tokens}{A sequence of ASM tokens for one row of the ASM file. This may be a row containing
(1) only a label
(2) a PC(*) assignment
(3) a variable assignment
(4) an assembler directive e.g. '.text', '.byte'
(5) an instruction
(6) any of the above preceded by a label}
}
\value{
single row data.frame containing information about the row
}
\description{
Determine the opcode, address mode etc for a sequence of tokens representing a single instruction
}
