% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{h2i}
\alias{h2i}
\title{Hex (character strings) to integer values}
\usage{
h2i(...)
}
\arguments{
\item{...}{can be a vector of values or multiple arguments}
}
\value{
vector of integers
}
\description{
Hex (character strings) to integer values
}
