% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TASS.R
\docType{data}
\name{TASS}
\alias{TASS}
\title{Class wrapping the 64TASS assembler for easier testing of my assembler}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
TASS
}
\description{
Class wrapping the 64TASS assembler for easier testing of my assembler
}
\examples{
\dontrun{
tass <- TASS$new("./asm/border.asm")
tass$dump_asm()
tass$compile()
tass$get_prg()
tass$get_asm()
tass$compile_and_run()
}
}
\keyword{datasets}
