% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TASS.R
\docType{data}
\name{TASS}
\alias{TASS}
\title{Class wrapping the 64TASS assembler for easier testing of the R64 assembler}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
TASS
}
\description{
Set `option(TASS_BIN=...)` and `option(X64_BIN=...)` to set the assembler
and emulator executable locations, or pass in as arguments
`TASS$new(tass_bin=..., x64_bin=...)`
}
\examples{
\dontrun{
tass <- TASS$new("./asm/border.asm", tass_bin = "~/bin/64tass", x64_bin = "~/bin/x64")
tass$dump_asm()
tass$compile()
tass$get_prg()
tass$get_asm()
tass$compile_and_run()
}

}
\keyword{datasets}
