% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TASS.R
\name{TASS}
\alias{TASS}
\title{Class wrapping the 64TASS assembler for easier testing of the R64 assembler}
\description{
Set `option(TASS_BIN=...)` and `option(X64_BIN=...)` to set the assembler
and emulator executable locations, or pass in as arguments
`TASS$new(tass_bin=..., x64_bin=...)`
}
\examples{
\dontrun{
tass <- TASS$new("./asm/border.asm", tass_bin = "~/bin/64tass", x64_bin = "~/bin/x64")
tass$dump_asm()
tass$compile()
tass$get_prg()
tass$get_asm()
tass$compile_and_run()
}

}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{TASS$new()}}
\item \href{#method-compile}{\code{TASS$compile()}}
\item \href{#method-run}{\code{TASS$run()}}
\item \href{#method-dump_asm}{\code{TASS$dump_asm()}}
\item \href{#method-get_asm}{\code{TASS$get_asm()}}
\item \href{#method-get_prg}{\code{TASS$get_prg()}}
\item \href{#method-compile_and_run}{\code{TASS$compile_and_run()}}
\item \href{#method-clone}{\code{TASS$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TASS$new(
  asm_filename = NULL,
  asm = NULL,
  prg_filename = NULL,
  tass_bin = "~/bin/64tass",
  x64_bin = "/usr/local/opt/vice/x64.app/Contents/MacOS/x64 -VICIIfilter 0  -silent"
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-compile"></a>}}
\subsection{Method \code{compile()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TASS$compile()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-run"></a>}}
\subsection{Method \code{run()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TASS$run(wait = FALSE, intern = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dump_asm"></a>}}
\subsection{Method \code{dump_asm()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TASS$dump_asm()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_asm"></a>}}
\subsection{Method \code{get_asm()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TASS$get_asm()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_prg"></a>}}
\subsection{Method \code{get_prg()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TASS$get_prg()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-compile_and_run"></a>}}
\subsection{Method \code{compile_and_run()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TASS$compile_and_run(wait = FALSE, intern = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TASS$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
